news_ph/pyproject.toml
[project]
name = "news-ph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dagster>=1.10.20",
    "dagster-dbt>=0.26.20",
    "dagster-duckdb>=0.26.20",
    "dagster-duckdb-pandas>=0.26.20",
    "dagster-postgres>=0.26.20",
    "dbt-duckdb>=1.9.3",
    "load-dotenv>=0.1.0",
    "pandas>=2.3.0",
    "tenacity>=9.1.2",
]
